requirements.txt
requests==2.31.0
python-dotenv==0.9.1
python-telegram-bot==13.14
